Move query validation logic to Editors/View component
