+++ src/components/Controllers/View/store.ts
@@ -1,7 +1,6 @@
 import * as React from 'react'
 import * as zustand from 'zustand'
 import create from 'zustand/vanilla'
-import { Parser, AST, Select } from 'node-sql-parser'
 import { assert } from 'ts-essentials'
 import { IFile, IViewError, ViewErrorLocation } from '../../../interfaces'
 import { Client } from '../../../client'
@@ -51,75 +50,20 @@ export function createStore(props: SqlProps) {
     makeQuery: async () => {
       const { client, view } = get()
       if (!view) return
-      const parser = new Parser()
-      let parsedSQL
 
       try {
-        parsedSQL = parser.astify(view.query)
+        const { table } = await client.tableQuery({ query: view.query })
+        set({ table })
+        set({ viewError: undefined })
       } catch (error) {
         const errorObj: IViewError = {
-          message: (error as ExceptionError).message,
-          location: ViewErrorLocation.Frontend,
+          message: 'Error response from Frictionless API',
+          location: ViewErrorLocation.Backend,
         }
         set({ viewError: errorObj })
       }
-      if (parsedSQL) {
-        const { tables } = get()
-        const { fields } = get()
-
-        const errors = checkExistingTablesAndFields(parsedSQL, tables, fields)
-        if (errors.length > 0) {
-          const errorObj: IViewError = {
-            message: errors.join(' '),
-            location: ViewErrorLocation.Frontend,
-          }
-          set({ viewError: errorObj })
-          return
-        }
-        try {
-          const { table } = await client.tableQuery({ query: view.query })
-          set({ table })
-          set({ viewError: undefined })
-        } catch (error) {
-          const errorObj: IViewError = {
-            message: 'Error response from Frictionless API',
-            location: ViewErrorLocation.Backend,
-          }
-          set({ viewError: errorObj })
-        }
-      }
     },
   }))
-}
-
-const checkExistingTablesAndFields = (
-  sqlAST: AST | AST[],
-  tables: string[] | undefined,
-  fields: IFieldItem[] | undefined
-) => {
-  const errors: string[] = []
-  const select: Select = sqlAST as Select
-
-  if (select !== null && select.from) {
-    for (const t of select.from) {
-      if (tables && tables.indexOf(t.table) < 0) {
-        errors.push(`Table "${t.table}" does nos exist.`)
-      }
-    }
-
-    if (select.columns) {
-      for (const c of select.columns) {
-        if (c.expr && c.expr.column) {
-          const column = c.expr.column
-          if (fields && fields.findIndex((f) => f.name === column) < 0) {
-            errors.push(`Field "${c.expr.column}" does nos exist.`)
-          }
-        }
-      }
-    }
-  }
-
-  return errors
 }
 
 export function useStore<R>(selector: (state: State) => R): R {
